requirements.txt
customtkinter
pillow
requests
prisma